+++ src/FennecLabs.Instrumentation/Output/OutputFormat.cs
@@ -1,0 +1,4 @@
+namespace FennecLabs.Instrumentation.Output
+{
+    public enum OutputFormat { Fxt, Json }
+}